Move Dashboard components to separate folders
Problem: it's getting too messy in page/dashboard.

Solution: move components to their own folders.
