+++ frontend/src/pages/dashboard/DashboardPage.tsx
@@ -3,11 +3,11 @@ import { useRecoilValue, waitForAll } from "recoil";
 import PageLayout from "../../components/PageLayout/PageLayout";
 import { CompoundSelector, ProjectSelector, TargetSelector } from "./EntitySelector";
 import "./DashboardPage.scss";
-import { ExperimentsTableSuspendable } from "./ExperimentsTable";
+import { ExperimentsTableSuspendable } from "./ExperimentsTable/ExperimentsTable";
 import { SuspenseSpinner } from "../../components/Spinner/SuspsenseSpinner";
 import cn from "../../utils/bemUtils";
 import { experimentsTableSizeAtom, selectedSubExperimentsColorAtom } from "../../store/atoms";
-import PlotlyChart from "./Plotting";
+import PlotlyChart from "./Plotting/Plotting";
 import { selectedSubExperimentsQuery } from "../../store/selectors";
 import { negateTableSize, SubExperimentWithMeasurements } from "../../store/types";
 import { isDefined, zip } from "../../utils/utils";
