+++ frontend/src/pages/dashboard/ExpandMinimizeButton/ExpandMinimizeButton.scss
