+++ frontend/src/pages/dashboard/Plotting/Plotting.scss
