+++ frontend/src/pages/dashboard/SubexperimentPlate/SubexperimentPlate.scss
